Formula for pango
Pango is a library for laying out and rendering of
text, with an emphasis on internationalization.
Pango can be used anywhere that text layout is
needed, though most of the work on Pango so far
has been done in the context of the GTK+ widget
toolkit. Pango forms the core of text and font
handling for GTK+-2.x.