Formula for cairo
Cairo is a 2D graphics library with support for
multiple output devices.

OS X includes the cairo library, but only with
rendering support for the X Server. This package
uses cairos quartz backend.

This allows applications using cairo to run as
native OS X applications, not using the X server.